+++ Specifications/Services/Services.csproj
@@ -9,6 +9,7 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.Reactive.Testing" Version="$(RxVersion)" />
+        
     </ItemGroup>
 
     <ItemGroup>
