+++ Specifications/Services/for_ServerStreamingEnumerable/when_enumerating/and_cancellation_is_requested.cs
@@ -3,19 +3,19 @@
 
 using System.Collections.Generic;
 using System.Threading;
-using Dolittle.SDK.Services.for_ServerStreamingMethodHandler.given;
+using Dolittle.SDK.Services.for_ServerStreamingEnumerable.given;
 using Dolittle.SDK.Services.given;
 using Machine.Specifications;
 
-namespace Dolittle.SDK.Services.for_ServerStreamingMethodHandler.when_aggregating_responses;
+namespace Dolittle.SDK.Services.for_ServerStreamingEnumerable.when_enumerating;
 
 public class and_cancellation_is_requested : given.all_dependencies
 {
-    static ServerStreamingMethodHandler<Message> handler;
+    static ServerStreamingEnumerable<Message> handler;
     static IEnumerable<Message> result;
     Establish context = () =>
     {
-        handler = create_handler(new FakeAsyncStreamReader<Message>(new Message(41), new Message(52)));
+        handler = create_enumerable(new FakeAsyncStreamReader<Message>(new Message(41), new Message(52)));
     };
 
     Because of = () => result = handler.AggregateResponses(new CancellationToken(true)).GetAwaiter().GetResult();
