+++ Specifications/Services/for_ServerStreamingEnumerable/when_enumerating/and_there_are_two_messages_in_response_stream.cs
@@ -2,16 +2,15 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using System.Threading;
-using Dolittle.SDK.Services.for_ServerStreamingMethodHandler.given;
+using Dolittle.SDK.Services.for_ServerStreamingEnumerable.given;
 using Dolittle.SDK.Services.given;
 using Machine.Specifications;
 
-namespace Dolittle.SDK.Services.for_ServerStreamingMethodHandler.when_aggregating_responses;
+namespace Dolittle.SDK.Services.for_ServerStreamingEnumerable.when_enumerating;
 
 public class and_there_are_two_messages_in_response_stream : given.all_dependencies
 {
-    static ServerStreamingMethodHandler<Message> handler;
+    static ServerStreamingEnumerable<Message> handler;
     static IEnumerable<Message> messages_in_response_stream;
     static IEnumerable<Message> result;
     Establish context = () =>
@@ -21,7 +20,7 @@ public class and_there_are_two_messages_in_response_stream : given.all_dependenc
             new Message(41),
             new Message(52)
         };
-        handler = create_handler(new FakeAsyncStreamReader<Message>(messages_in_response_stream));
+        handler = create_enumerable(new FakeAsyncStreamReader<Message>(messages_in_response_stream));
     };
 
     Because of = () => result = handler.AggregateResponses().GetAwaiter().GetResult();
